refactor(file-structure): introducing _component folders for non re-usable components
