+++ app/contact/page.tsx
@@ -1,5 +1,5 @@
 // COMPONENTS
-import { ContactForm } from "@/components/contact-form";
+import { ContactForm } from "./_components/contact-form";
 
 
 export default async function ContactPage() {
